requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
twilio==8.10.0
google-generativeai==0.3.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.1.0
